moved database to new dir, renamed files and functions, updated help
